airflow/requirements.txt
apache-airflow-providers-common-sql
pyarrow
pyspark
azure-storage-file-datalake
azure-storage-blob
pandas